Renamed Smacht, added deploy and GPS time set.
Renamed the Smacht module to be SGS (Sailboat Guidance System in
honour of the Apollo Guidance System), added a deploy rule to the
top-level Makefile to push the code to the boat, and also added
a gps_time command which reads the time from the GPS and sets the
system time. This is used because the ALIX boards don't have an
RTC chip. This command gets executed early during the boot process
so that the system time is correct as soon as possible.